requirements.txt
pandas>=2.2.2
numpy>=1.26.4
pyyaml>=6.0.2
tqdm>=4.66.4
pytest>=8.1.1
